+++ src/rextio_device_cuda/__init__.py
@@ -10,4 +10,3 @@ __all__ = [
     "__version__",
     "provider",
 ]
-